tpg/Transparent_PNG_Generator/requirements.txt
flask==2.2.5
flask-wtf==1.1.1
flask-sqlalchemy==2.4.4  # バージョンを 2.4.4 に変更
pillow==9.5.0
python-magic==0.4.27
gunicorn==21.2.0
flask-limiter==3.3.1
werkzeug==2.2.3
email-validator==2.0.0
requests==2.29.0
